+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@
 - GitHub workflows:
   - Renamed to more meaningful names.
   - Added `darwin_build_cpu_pjrt.yaml`: workflow to automatically build the Darwin CPU PJRT plugin.
+  - Remove unnecessary `apt install` in linux tests.
 - Package `xlabuilder`:
   - Added `-std=gnu11` to `CFLAGS` to avoid incompatibility introduced in `c23`. 
   - Temporary fix, as `xlabuilder` is deprecated and will be removed in the future.