+++ docs/CHANGELOG.md
@@ -6,6 +6,7 @@
 - Package `pjrt`:
   - Renamed `Client.Devices` to `Client.AllDevices`.
   - Added `Client.NumDevices` to get the number of addressable devices.
+  - Removed `xlabuilder` dependency on tests, replaced by `stablehlo` package instead.
 
 # v0.8.5 2025/10/27: Fixed `xlabuilder` for newer C compilers.